Add files via upload
The images and css and js codes are supposed to be in a file dedicated to each type of file such as for example js. So in the php and html code files, you need to remove the "file/" that is after href command in these codes (htmls' and the php).